requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
django-bootstrap-calendar==0.1.1
django-multiselectfield==0.1.12
django-phonenumber-field==7.0.2
django-tempus-dominus==5.1.2.17
django-twilio==0.14.3.1
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
Events==0.4
fonttools==4.41.1
freetype-py==2.3.0
frozenlist==1.3.3
html5lib==1.1
icalendar==5.0.4
idna==3.4
jsmin==3.0.1
lxml==4.9.2
msgpack==1.0.5
multidict==6.0.4
mysql-connector-python==8.0.32
mysqlclient==2.1.1
oauthlib==3.2.2
openpyxl==3.1.2
oscrypto==1.3.0
pdfkit==1.0.0
phonenumbers==8.13.9
Pillow==9.5.0
pycairo==1.23.0
pycparser==2.21
pydyf==0.7.0
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
PyJWT==2.7.0
PyNaCl==1.5.0
pypdf==3.12.0
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
redis==5.0.0
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
rlPyCairo==0.2.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
telnyx==2.0.0
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.2
tzlocal==5.0.1
uk-postcode-utils==1.1
uritools==4.0.1
urllib3==2.0.3
weasyprint==59.0
webencodings==0.5.1
whitenoise==6.5.0
wkhtmltopdf==0.2
xhtml2pdf==0.2.11
yarl==1.8.2
zopfli==0.2.2
